+++ src/Utils/index.ts
@@ -12,3 +12,5 @@ export * from './lt-hash'
 export * from './auth-utils'
 export * from './legacy-msgs'
 export * from './baileys-event-stream'
+export * from './use-single-file-auth-state'
+export * from './use-multi-file-auth-state'
